♻️ Use Concurrency instead of Async class
